% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nbcoherence.R
\name{compare_nbcoherence}
\alias{compare_nbcoherence}
\title{Plot and compare narrow-band coherence between two channels}
\usage{
compare_nbcoherence(
  x = NULL,
  y = NULL,
  target_rhythm = "alpha",
  max_lags_plot = 40,
  max_lags_calculate = 40,
  max_plots = 3,
  fs = 100,
  N = 600
)
}
\description{
Plot and compare narrow-band coherence between two channels
}
\examples{
\dontrun{
 compare_nbcoherence(max_lags_plot=200, max_lags_calculate=200)
}
}
\keyword{EEG}
