% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rhythms.R
\name{extract_rhythms}
\alias{extract_rhythms}
\title{Extract EEG brain rhythms from a time series}
\usage{
extract_rhythms(x, fs, rhythms = NULL, print_filters = TRUE)
}
\description{
Extract EEG brain rhythms from a time series
}
\examples{
\dontrun{
  simulate_unstable_oscillator(100, 10, 4, 30) # Sample 100 points of an 10Hz oscillator with $tau=4$ sampled at 30Hz
 x <- simulate_unstable_oscillator(1000, 1, 1, 200);
 y <- extract_rhythms(x, 200)
 plot(y$delta); lines(y$theta, col="red"); lines(x, col="blue")
}
}
\keyword{EEG}
